Agregue la clase aplicacion, todavia no la modifique para que funcione

+++ src/turismo_en_tierra_media/Atraccion.java
@@ -4,11 +4,11 @@ public class Atraccion {
 
 	private String nombre;
 	private int valor;
-	private float tiempoDeUso;
+	private double tiempoDeUso;
 	private int usosDisponibles;
 	private TipoDeAtraccion tipo;
 	
-	public Atraccion(String nombre, int valor, float tiempoDeUso, int usosMaximos, TipoDeAtraccion tipo) {
+	public Atraccion(String nombre, int valor, double tiempoDeUso, int usosMaximos, TipoDeAtraccion tipo) {
 
 		this.nombre=Validacion.validar(nombre);
 		this.valor=Validacion.validar(valor);
@@ -66,5 +66,7 @@ public class Atraccion {
 		this.usosDisponibles-=1;
 		unUsuario.agregarAtraccion(this);
 	}
+	
+	/////////////////////
     
 }
